implement sender handshake with peer
Sender uses the transit key to generate a hexid and sends a handshake
message. Receiver sends back a receiver handshake message. Sender
checks if the received message matches with its own generated
message and if so, it sends "go\n", else it sends "nevermind\n".

Successfully handshakes with the python client.