WIP: handle transit ack message
After the receiver receives the file as records and decrypts and writes it into the disk
successfully, it sends an ack back to the sender, which is of this form:

{"ack": "ok", "sha256": "e4f1684a5375ebf7f1dcde02a66026f937a8c6195adf31813ef21b3ccadfb11f"}

This is sent as an encrypted record. So, we read length and read length sized record
and try to decrypt it and then deserialize the json message and check the 'ack' key
and 'sha256' key.