move Helper out of the library and refactor options parser into its own module
Helper defines just one function that prints a help text on what is to be typed
on the other side. This is ideally in the realms of the top level sendMessage and
sendFile function. Instead of calling it from these functions, we parameterize
this helper function and pass it as an argument, so a different helper text
can be given by another user of the library.

We also move out the command line options parser into its own module.