Calculate sha256 of encrypted records and check against recipient's sha256 sum
As part of the final Ack from the receiver, we get the sha256sum of
the encrypted records from the receiver. At the sending side, we
calculate the sha256 sum and compare it at the end with the one received
from the recipient and flag an error if there is a mismatch.