Use the newly exported deriveKey function from wormhole library
We had a proposed change to wormhole library where we exposed sessionKey
and deriveKey function. However, exposing sessionKey is probably not a good
idea and instead we are changing the proposed change to wormhole library
by exposing a version of deriveKey which takes the EncryptedConnection
as the input. We don't expose constructors of EncryptedConenction, so it
is just an opaque type that represents the encrypted connection established
by the wormhole protocol with the peer.