WIP: file receive  support
Start a listener and concurrently start listening while
attempting to connect to the peer hints.

File transfer between haskell/python clients work.

Add some simple resource leak fixes like closing the sockets. This
does not guarantee cleaning up the resources in the case of exceptions,
but this is the bare minimum to start with.