split receiveRecords into actually receving decrypted records and file write
The previous version of receiveRecords did too much. It received encrypted
records from the network, decrypted and wrote it, record by record, to the
disk. Instead, we now do receive and decrypt alone in receiveRecords function,
keep the entire blocks representing the sent file in memory, calculate sha256
and then write it into the disk.

While this may be inefficient for large files, this seem like a very logical
thing to do from splitting responsibilities of various functions.

At a later point, we could think of using conduit or pipes.