Address the review comments from @exarkun
The sourceFile does the right thing by sending a chunk of size
`defaultChunkSize' defined in Data.ByteString.Internal.Lazy
which is currently defined as 32k. So, we don't really need to do
our own chunking.

The pipeline gets rid of `takeC`. `takeC would take n chunks from
the upstream (in our case, it would stop the pipeline after sending
n times 32KBytes). `takeC` "worked" because we tested with shorter
files.

The pipeline now returns a Text type which is a hex representation of
the Sha256 hash of the input. The `Show' instance for Digest does that
magic and gives us the hex representation.

A few other renaming of variables as well to avoid cryptic names.