get rid of allocateTcpPort
allocateTcpPort suffers from a race condition, as mentioned in #18
and the cited tahoe-lafs ticket 2787. We replace allocateTcpPort
with a function that starts a tcp listener socket, which we query
to find the port. The socket returned by the listener is later used
to start an async thread that is waiting for a connection.