each peer should connect to a set of combined hints
In our initial implementation, we only connected to the hints we got from
our peer. However, imagine a situation where one peer is configured to use
a transit server and the other peer is configured to use another. If the peers
only talk to the hints we got from the peer, we would just deadlock at the
handshake. Instead, we should combine all the hints (the peers could as well
be configured to use the same transit server) as a set and try to connect to
them.