We now work in a ReaderT context.
The main "app" monad is a transformer:

 ReaderT Env (ExceptT Error IO a)

The `ExceptT e IO a' is a pseudonum for `IO (Either e a)' which we
had been using all throughout. Env provides a read-only environment
or a configuration for the entire application.

Next step would be to define a newtype for this monad transformer
stack.