bug fix: hangs on sending file to a remote peer via relay
While connecting to a set of hosts defined by the hints, we were
listing the locally listening server as well to the list of hosts
that the local client is connecting to, which resulted in client
getting connected to the locally listening server and so tryToConnect
call succeeds but nothing much happens after that. The solution is
to maintain only a list of peer hint and our relay hint as the potential
list of peers to connect to.