Ignore the base path of the filename in the Offer message while writing the file
While the sender takes care of sending only the file name, a malicious client
could send us a file path and let us overwrite a file with malicious content. So
instead of relying on the sender to do the right thing, the receiver extracts the
filename alone and writes into it in the current directory.

Closes #27.