bug fix: do not ignore the path returned by the senderOfferExchange
senderOfferExchange returns the zip of the directory to be transferred
or the file path for a file transfer. This needs to be passed to the
pipeline that takes the file and sends it across by creating chunks,
encrypting each record and so on.. For directory transfer, this was
broken as the returned path was being ignored. For file transfer it
worked because the original file path is available as a parameter
passed to the function.