Fix Ord instance for ConnectionHint and Ipv6 support.
Fix a subtle bug where an Ord instance for ConnectionHint casused a
hint to be dropped. While comparing two Direct hints, if they are
deemed equal, then while adding them to a set, one of them get dropped
out. This happens when an interface has an ipv4 and ipv6 listening
socket and we enumerate interfaces and send hints to the other side.