resume an interrupted download
A proof of concept implementation (and a slight protocol change) for a
scenario where a download is interrupted for various reasons like
network interruption or user aborting the program and then resuming
the download from the point where it ended in the last failed attempt
instead of starting all over again.

After the offer message, the recipient sends an "ack" message which
currently consist of a OK message. Along with it, we also send an
offset (if the file is not downloaded earlier, an offset of 0 is
sent. If it is partially downloaded, the current size at the recipient
side is the offset that is sent back to the sender). The sender then
starts from that offset and recipient decrypts and appends to the
already downloaded parts.